+++ src/titiler/xarray/titiler/xarray/extensions.py
@@ -2,14 +2,13 @@
 
 import warnings
 from collections.abc import Callable
-from typing import Annotated
+from typing import Annotated, TypedDict
 
 import xarray
 from attrs import define
 from fastapi import Depends, Query
 from rio_tiler.constants import WGS84_CRS
 from starlette.responses import HTMLResponse
-from typing_extensions import TypedDict
 
 from titiler.core.dependencies import DefaultDependency
 from titiler.core.factory import FactoryExtension